requirements.txt
pypdf
pytesseract
Pillow
transformers
torch
numpy
